requirements.txt
logging
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
txtai
torch
nltk
fastapi
uvicorn
requests
aiohttp
openai
protobuf
numpy
tqdm
pydantic
qdrant-client 
sentence-transformers
pymongo
langchain
langchain-core
langchain-community
langchain-openai
langchain-qdrant
sentence-transformers
fastembed